refactor: remove separate builder test and incorporate the functionality into existing test files
